Adding fix for timeout issue and file checks for proper C file (#27)
* Adding fix for timeout issue and file checks for proper C file

* Adding file size checks while generating output

* Add checks for output format, submission watcher

* Cleaning code and adding proper comments

* Disable execution if clang checks fail

* Adding fix for catching implicit system function calls using clang tool

* Add option to close submissions, make delete button available after close, move submission limit to problem model

* Adding error message, filename fix

* Fix trailing whitespace for ascii check

* Option to disable evaluation, separate clang check failed submissions

* Fix undefined behaviour bug, C file check, verdict type subdirectories in zip, skip logs for CE

+++ content/compile_and_test.py
@@ -16,10 +16,18 @@ parser.add_argument('--submission_config', type=str,
                             ...""")
 args = parser.parse_args()
 
+def is_binary(fileName):
+    try:
+        with open(fileName) as check_file:  # try open file in text mode
+            check_file.read()
+            return False
+    except:  # if fail then file is non-text (binary)
+        return True
+
 with open(args.submission_config) as f:
     sub_info = [x[:-1] for x in f.readlines()]
 
-#print(sub_info)
+print(sub_info)
 
 # Retain the first 2 lines alone
 subprocess.call(['rm', args.submission_config])
@@ -28,33 +36,60 @@ with open(args.submission_config, "w") as stat_file:
 
 
 # Run clangjudge only for c/c++ submissions
+clangjudge_flag = 0
 if (sub_info[2]==".c" or sub_info[2]==".cpp") and sub_info[3]:
-    try:
-        subprocess.check_output(['./judgeClangTool/clangjudge', sub_info[3],
+   
+    isFileBinary=is_binary('submissions/submission_{}{}'.format(sub_info[1], sub_info[2]))
+    #print(isFileBinary)
+
+    if(isFileBinary):
+        #print("binary found!")
+        pass
+    
+    else:
+        try:
+            subprocess.check_output(['./judgeClangTool/clangjudge', sub_info[3],
                                 'submissions/submission_{}{}'.format(sub_info[1], sub_info[2])])
-    except subprocess.CalledProcessError as e:
-        error_msg = str(e.output.decode('utf-8'))
-        clangtool_log_file = 'sub_clangjudge_{}.log'.format(sub_info[1])
-        with open('tmp/' + clangtool_log_file, "w") as log_file:
-            log_file.write(error_msg)
+        except subprocess.CalledProcessError as e:
+            error_msg = str(e.output.decode('utf-8'))
+            clangtool_log_file = 'sub_clangjudge_{}.log'.format(sub_info[1])
+            with open('tmp/' + clangtool_log_file, "w") as log_file:
+                log_file.write(error_msg)
+            clangjudge_flag = 1
+            with open(args.submission_config, "a") as stat_file:
+                log_file_name = 'sub_run_{}_{}.log'.format(sub_info[1], sub_info[6])
+
+                with open('tmp/' + log_file_name, "w") as log_file:
+                    log_file.write("Clang Checks failed. Program uses function that is not allowed.")
+
+                stat_file.write("{} {} 0 0 {}\n".format(sub_info[6], 'CE2', log_file_name))
+                for testcase_id in sub_info[7:]:
+                    stat_file.write("{}\n".format(testcase_id))
 
 # First compile
-try:
-    subprocess.check_output(['./main_compiler.sh', sub_info[0],
-                             'submission_{}{}'.format(sub_info[1], sub_info[2])],
-                            stderr=subprocess.STDOUT)
-except subprocess.CalledProcessError as e:  # If compilation fails, end this script here
-    error_msg = str(e.output.decode('utf-8'))
-    with open(args.submission_config, "a") as stat_file:
-        for testcase_id in sub_info[6:]:
-            log_file_name = 'sub_run_{}_{}.log'.format(sub_info[1], testcase_id)
+if clangjudge_flag == 0:
+    try:
+        subprocess.check_output(['./main_compiler.sh', sub_info[0],
+                                'submission_{}{}'.format(sub_info[1], sub_info[2])],
+                                stderr=subprocess.STDOUT)
+    except subprocess.CalledProcessError as e:  # If compilation fails, end this script here
+        if isFileBinary or e.returncode == 2:
+            error_msg = "Submitted file has invalid format."
+            verdict_type = 'CE1'
+        else:
+            error_msg = str(e.output.decode('utf-8'))
+            verdict_type = 'CE0'
+
+        with open(args.submission_config, "a") as stat_file:
+            log_file_name = 'sub_run_{}_{}.log'.format(sub_info[1], sub_info[6])
 
             with open('tmp/' + log_file_name, "w") as log_file:
                 log_file.write(error_msg)
 
             stat_file.write("{} {} 0 0 {}\n"
-                            .format(testcase_id,
-                                    'CE' if e.returncode == 1 else 'NA', log_file_name))
-else:
-    subprocess.call(['./main_tester.sh'] + sub_info[0:2] + sub_info[4:])  # run tests
-    subprocess.call(['rm', 'submissions/submission_{}'.format(sub_info[1])])  # remove executable
+                            .format(sub_info[6], verdict_type, log_file_name))
+            for testcase_id in sub_info[7:]:
+                stat_file.write("{}\n".format(testcase_id))
+    else:
+        subprocess.call(['./main_tester.sh'] + sub_info[0:2] + sub_info[4:])  # run tests
+        subprocess.call(['rm', 'submissions/submission_{}'.format(sub_info[1])])  # remove executable
